% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict}
\alias{dict}
\title{Scorecard data dictionary}
\format{An object of class \code{data.frame} with 1976 rows and 6 columns.}
\source{
\url{"https://collegescorecard.ed.gov/assets/CollegeScorecardDataDictionary.xlsx"}
}
\usage{
dict
}
\description{
Scorecard data dictionary
}
\keyword{datasets}

