% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{filter_}
\alias{reexports}
\alias{interp}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter_}}}

  \item{lazyeval}{\code{\link[lazyeval]{interp}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

