% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_dict_df}
\alias{sc_dict_df}
\title{Scorecard data dictionary}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1975 rows and 7 columns.}
\source{
\url{"https://collegescorecard.ed.gov/assets/CollegeScorecardDataDictionary.xlsx"}
}
\usage{
sc_dict_df
}
\description{
Scorecard data dictionary
}
\keyword{datasets}

