% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{sc_get}
\alias{sc_get}
\title{Get scorecard data.}
\usage{
sc_get(sccall, api_key, debug = FALSE, print_key_debug = FALSE)
}
\arguments{
\item{sccall}{Current list of parameters carried forward from prior
functions in the chain (ignore)}

\item{api_key}{Personal API key requested from
\url{https://api.data.gov/signup} stored in a string. If you
first set your key using \code{sc_key}, then you may omit this
parameter. A key set here will take precedence over any set in
the environment (DATAGOV_API_KEY).}

\item{debug}{Set to true to print and return API call (URL string)
rather than make actual request. Should only be used when
debugging calls.}

\item{print_key_debug}{Only used when \code{debug == TRUE}. Default
masks the \code{api_key} value. Set to \code{TRUE} to print the
full API call string with the \code{api_key} unmasked.}
}
\description{
This function gets the College Scorecard data by compiling and
converting all the previous piped output into a single URL string
that is used to get the data.
}
\section{Obtain a key}{

To obtain an API key, visit \url{https://api.data.gov/signup}
}

\examples{
\dontrun{
sc_get('<API KEY IN STRING>')
key <- '<API KEY IN STRING>'
sc_get(key)
}

}
