% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{sc_init}
\alias{sc_init}
\title{Initialize chained request.}
\usage{
sc_init(dfvars = FALSE)
}
\arguments{
\item{dfvars}{Set to \code{TRUE} if you would rather use the
developer-friendly variable names used in actual API call.}
}
\description{
This function initializes the data request. It should always be
the first in the series of piped functions.
}
\examples{
\dontrun{
sc_init()
sc_init(dfvars = TRUE)
}
}
