% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{sc_key}
\alias{sc_key}
\title{Store Data.gov API key in system environment.}
\usage{
sc_key(api_key)
}
\arguments{
\item{api_key}{Personal API key requested from
\url{https://api.data.gov/signup} stored in a string.}
}
\description{
This function stores your data.gov API key in the system environment
so that you only have to load it once at the start of the session.
If you set your key using \code{sc_key}, then you may omit
\code{api_key} parameter in the \code{sc_get} function.
}
\section{Obtain a key}{

To obtain an API key, visit \url{https://api.data.gov/signup}.
}

\examples{
\dontrun{
sc_key('<API KEY IN STRING>')
}

}
