% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscorecard.R
\docType{package}
\name{rscorecard}
\alias{rscorecard}
\alias{rscorecard-package}
\title{rscorecard: A method to download College Scorecard data.}
\description{
The rscorecard package provides a series of piped functions (a la
\code{\link[dplyr]{dpylr}}) to facilitate downloading Department of
Education College Scorecard data.  In reality it is simply a method
for converting idiomatic R code into a properly formatted URL
string that is then queried. This package requires an API key,
which can be requested at
\url{https://api.data.gov/signup/}.
}

